refactor(saves): decompose sync_engine into matrix / devices / rollback (#495)
services/saves/sync_engine/engine.py was 1203 LOC, ~2x the [ours]
god-class threshold. Splits into focused internal sub-modules:

- engine.py — SyncEngine entry point + per-rom lock dispatcher
- matrix.py — newest-wins matrix executor + per-file sync I/O dispatch
- devices.py — RomM device registration
- rollback.py — conflict-resolution rollback flow (use_server / keep_local)

Sub-modules are internal helpers; SyncEngine instantiates them during
__init__ from the existing ctor kwargs and exposes thin delegate
methods so peer save services (status, versions, slots) and tests
keep working through the same SyncEngine surface. Public surface,
behavior, and test suite unchanged.

Closes #495
